stress-sctp: don't unlink on an address that is NULL
addr may not be set, so don't unlink on a deferenced cast from it

+++ stress-sctp.c
@@ -183,7 +183,7 @@ static int stress_sctp_server(
 	int fd, status;
 	int so_reuseaddr = 1;
 	socklen_t addr_len = 0;
-	struct sockaddr *addr;
+	struct sockaddr *addr = NULL;
 	uint64_t msgs = 0;
 	int rc = EXIT_SUCCESS;
 
@@ -257,7 +257,7 @@ die_close:
 	(void)close(fd);
 die:
 #if defined(AF_UNIX)
-	if (opt_sctp_domain == AF_UNIX) {
+	if (addr && opt_sctp_domain == AF_UNIX) {
 		struct sockaddr_un *addr_un = (struct sockaddr_un *)addr;
 		(void)unlink(addr_un->sun_path);
 	}
